assistant/requirements.txt
fastapi==0.95.1
pydantic==1.10.7
uvicorn==0.22.0
uvloop==0.17.0
aiohttp==3.8.4
orjson==3.8.12
redis[hiredis]==4.5.5
fastapi-cache2[redis]==0.2.1

